// web/lib/api/models/availableSlotDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Reservae API
 * API para gerenciamento de espaços, disponibilidades e reservas.
 * OpenAPI spec version: v1
 */

export interface AvailableSlotDto {
  /** @nullable */
  bookableSlotId?: number | null;
  /** @nullable */
  availabilityRuleId?: number | null;
  spaceId?: number;
  startsAt?: string;
  endsAt?: string;
  pricePerSpot?: number;
  capacity?: number;
  reservedQuantity?: number;
  availableQuantity?: number;
  isVirtual?: boolean;
}